Add modules collector tests
