+++ src/js/tools/collectModules.ts
@@ -1,6 +1,8 @@
-import { existsSync, readFileSync, writeFileSync } from 'fs';
-import { dirname, join, resolve } from 'path';
+import { readFileSync, writeFileSync } from 'fs';
+import { resolve } from 'path';
 import { argv } from 'process';
+
+import ModulesCollector from './ModulesCollector';
 
 const sourceMapPath: string | undefined = argv[2]; // eslint-disable-line @typescript-eslint/no-unsafe-member-access
 const outputModulesPath: string | undefined = argv[3]; // eslint-disable-line @typescript-eslint/no-unsafe-member-access
@@ -14,6 +16,9 @@ if (!sourceMapPath) {
 if (!outputModulesPath) {
   throw new Error('Second argument `modules-output-path` is missing!');
 }
+if (modulesPaths.length === 0) {
+  throw new Error('Third argument `modules-paths` is missing!');
+}
 
 const map: { sources?: string[] } = JSON.parse(readFileSync(sourceMapPath, 'utf8'));
 if (!map.sources) {
@@ -21,56 +26,6 @@ if (!map.sources) {
 }
 
 const sources: string[] = map.sources;
-const modules = collectModules(sources, modulesPaths);
+const modules = ModulesCollector.collect(sources, modulesPaths);
 
 writeFileSync(outputModulesPath, JSON.stringify(modules, null, 2));
-
-function collectModules(sources: string[], modulesPaths: string[]): Record<string, string> {
-  const infos: Record<string, string> = {};
-  const seen: Record<string, true> = {};
-
-  sources.forEach((path: string) => {
-    let dir = path; // included source file path
-
-    /** Traverse directories upward in the search of all package.json files */
-    const upDirSearch = (): void => {
-      const parentDir = dir;
-      dir = dirname(parentDir);
-
-      const upDir = dirname(resolve(dir));
-      if (
-        !dir ||
-        parentDir === dir ||
-        seen[dir]
-      ) {
-        return;
-      }
-      seen[dir] = true;
-
-      const pkgPath = join(dir, 'package.json');
-      if (!existsSync(pkgPath)
-        ||!modulesPaths.includes(upDir)) {
-        // fast-forward if the package.json doesn't exist
-        // or the dir is not direct child of the modules path
-        return upDirSearch();
-      }
-
-      try {
-        const info: { name?: string, version?: string } = JSON.parse(readFileSync(pkgPath, 'utf8'));
-        if (info.name && info.version) {
-          infos[info.name] = info.version;
-        } else if (info.name) {
-          infos[info.name] = 'unknown';
-        }
-      } catch (_oO) {
-        // do-nothing
-      }
-
-      return upDirSearch(); // processed package.json file, continue up search
-    };
-
-    upDirSearch();
-  });
-
-  return infos;
-}
