+++ test/tools/fixtures/root-module/modules/module1/modules/not-collected/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "not-collected-module"
+}
