+++ test/tools/fixtures/root-module/modules/module1/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "module-1",
+  "version": "module-1-version"
+}
