+++ test/tools/fixtures/root-module/modules/module2/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "module-2"
+}
