+++ test/tools/fixtures/root-module/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "root-module",
+  "version": "root-module-version"
+}
